+++ src/widgets/DetailPanel/DetailPanelContent.tsx
@@ -3,10 +3,10 @@ import Button from "@/components/Button";
 import Slider from "@/components/Slider/Slider";
 import ToggleSwitch from "@/components/ToggleSwtich/ToggleSwtich";
 import { FieldErrors, FormValues } from "@/hooks/useRendererState";
+import { useTutorialState } from "@/hooks/useTutorialState";
 import clsx from "clsx";
 import React from "react";
 import { useConfirmation } from "../Confirmation/Confirmation";
-import { useTutorialState } from "@/hooks/useTutorialState";
 
 export interface DetailPanelProps {
   handleChange: (